Enabled type-specific behavior on contracts
Changed ContractRewriterBase.GetConditionExpression to take a TypeInfo
parameter instead of the typeName string.  This allows different code
generation based on the type the contract is applied to.

Changed basic comparisons and NonDefault behavior for nullable types.
If attribute is applied to T?, comparison is performed against
default(T) not default(T?) which is always null.

Added null check to NonEmpty if type is a reference type.